Fixed broken link to FractalCount plugin
Original plugin URL is broken, hence I replaced the link with my own copy in my Google Drive. It is a .class file, so it might be flagged as a virus by your antivirus. If in doubt, go to the original plugin github repo (https://github.com/perchrh/ImageJFractalDimension) and compile it.